feat(mcp): 将 McpClient 与 factory 迁移到 dasclaw_mcp（PR 6-γ）
将 ironclaw 中 1475 行的 `tools/mcp/client.rs`（McpClient 协议核心）和
222 行的 `tools/mcp/factory.rs`（按配置创建客户端）迁移到 host-agnostic 的
`dasclaw_mcp` crate。把 ironclaw 特有的 `Tool` trait 适配层抽离到新的
`desktop-client/ironclaw/src/tools/mcp/client_tool.rs`，通过
`RefreshAccessTokenFn` 注入钩子让 dasclaw_mcp 在不知道 ironclaw
OAuth 代理与 `JobContext` 的前提下，仍能走完整的刷新流程。

关键改动：

* 新增 `dasclaw_mcp::client::McpClient`（1331 行）：协议、状态机、token
  刷新调度。原 `refresh_access_token(...)` 直调换成
  `self.refresh_access_token_via_hook(...)`，未注入钩子时回退到
  `crate::auth::refresh_access_token_direct`（RFC 6749 直连路径）。
* 新增 `dasclaw_mcp::factory::create_client_from_config`：纯转 transport
  分发，依赖 `dasclaw_runtime::secrets::SecretsStore`。
* 新增 `pub type RefreshAccessTokenFn`：host 注入的刷新闭包，
  `Arc<dyn Fn(config, secrets, user_id) -> BoxFuture<Result<AccessToken,_>>>`。
* 新增 `McpClient::with_refresh_fn` builder；`Clone` 实现保留 `refresh_fn`
  以确保 client 副本仍持有 host 钩子。
* `tools/mcp/client.rs` 缩为 12 行 shim，`pub use dasclaw_mcp::client::McpClient`。
* `tools/mcp/factory.rs` 缩为 40 行包装，在内部 factory 返回后调用
  `.with_refresh_fn(ironclaw_refresh_fn())` 装上 OAuth 代理钩子。
* 新增 `tools/mcp/client_tool.rs`（245 行）：
  - `McpToolWrapper` + `impl Tool` —— 桥接 ironclaw 的 `Tool` trait
  - `create_tools_for(Arc<McpClient>)` —— 替代原 `McpClient::create_tools()`
  - `ironclaw_refresh_fn()` —— 包装 `tools::mcp::auth::refresh_access_token`
    走 OAuth 代理路径
  - `strip_top_level_nulls` —— 防御性的 LLM 参数清洗
* 调用点更新：`app.rs` 与 `extensions/manager.rs` 不再调
  `client.create_tools()`，改为先把 client 装进 Arc，再调
  `client_tool::create_tools_for(client_arc.clone())`，最后用同一个
  Arc 入注册表，避免重复包装。

验证：

* `cargo check -p dasclaw_mcp --tests`：clean
* `cargo nextest run -p dasclaw_mcp`：196/196 PASS（较 PR 6-β 的 165 +31）
* `cargo check -p dasclaw --tests`：clean（仅余 PR 6-β 留下的 shim 未用警告）
* `cargo nextest run -p dasclaw --lib -E 'test(mcp)'`：28/28 PASS
* `cargo clippy --no-deps -p dasclaw_mcp --all-targets -- -D warnings`：clean
* `cargo fmt --all`
* `python3.12 scripts/check_no_panics.py --base origin/xClaw`：OK

非目标：

* 不动 `tools/mcp/auth.rs`（仍持有 ironclaw 特有的 OAuth 代理路由逻辑，
  由 `ironclaw_refresh_fn` 包装注入）
* 不动 `tools/mcp/session.rs` / `tools/mcp/process.rs`（已在前几个 PR
  迁移到 dasclaw_mcp）
* 不重构刷新顺序、不引入新策略开关、不改 OAuth 代理协议

Closes #666
Refs #661
